farmer/requirements.txt
wget
sklearn
pandas
seaborn
matplotlib
dvc
google-api-python-client
dask[dataframe]
gcsfs
google-cloud
google-cloud-storage